Restructure: * Join backend + schme => type.dart
